correcting datatypes for postgres: int8, float8

+++ creating_sql_query_table_from_csv.py
@@ -20,8 +20,8 @@ for i, c in enumerate(df.columns):
 
 translating_dtypes_to_sqldatatypes = {
         'object': 'varchar',
-        'float64': 'NUMERIC',
-        'int64': 'NUMERIC',
+        'float64': 'float8', # 15 decimal digits precision
+        'int64': 'int8',
 }
